Extract utility functions from STF CLI to be re-used
Functions are used in main.rs of client and future commands
that will be added (e.g. place order)
